feat:Implement authorization

+++ blog_project/posts/views.py
@@ -20,8 +20,10 @@ def post_new(request):
     if request.method == 'POST':
         form = PostForm(request.POST)
         if form.is_valid():
-            form=form.save()
-            return redirect('post_detail',pk=form.pk)
+            post=form.save(commit = False)
+            post.author = request.user
+            post.save()
+            return redirect('post_detail',pk=post.pk)
     else:
         form = PostForm()
     return render(request,'posts/post_edit.html',{'form':form})
@@ -30,6 +32,8 @@ def post_new(request):
 def post_edit(request,pk):
     post = get_object_or_404(Post,pk=pk)
 
+    if post.author != request.user:
+        return redirect('post_list')
     if request.method=='POST':
         form = PostForm(request.POST,instance=post)
         if form.is_valid():
@@ -42,6 +46,10 @@ def post_edit(request,pk):
 @login_required
 def post_delete(request,pk):
     post = get_object_or_404(Post,pk=pk)
+
+    if post.author != request.user:
+        return redirect('post_list')
+
     if request.method == 'POST':
         post.delete()
         return redirect('post_list')
